Simplify and reduce scope to lkru only - Support only linux-keyring-utility 0.2.0+ - Strip unnecessary complexity
